merge to v1.2.0 code
\#2481

While reviewing the file joiner implementation, it was found that the code allows for an
unbounded recursive call.

\#1671

Previously, the pipeline builder, when it reached EOF after draining the request, compared the data length argument with the amount of data read. Since EOF signal is sent when all of the data is read gracefully, the amount of data read will always be equal to the length of the request, so this check was unnecessary, and removed with this PR.

\#1501

Removing collections package and API cleanup

\#1597

It is not a completely clean solution, as currently, it returns all Bee's custom headers across all endpoints, while a better solution would be to return only those that the endpoint indeed supports. But that would require quite a more complex support structure in the code that I was not really keen to introduce. I think this will make the work done, but it will require having it in mind when new headers are added to propagate them in here as well.

\#1807

remove state-store close call on nil reference

\#2380

The fix added earlier to fix the nodeType information on root nodes actually overrides the values in the case of other nodes, resulting in incorrect nodeType information on Lookup. Due to this, we will not process the node returned as a valid path and hence return 404 although we have the reference.

\#2487

improve localstore getmulti metrics

\#2489

SWA-01-012 WP2: Potential directory traversal in download API (Medium)

While reviewing the API implementation, it was found that the file download API possibly a directory traversal issue

It can be observed that the filename provided in the file’s metadata is directly returned in filename part of the Content-Disposition header. An attacker might, however, provide names such as ../../../../etc/passwd or similar, which could subsequently be accidentally used by client-code. In order to address this issue, it is recommended to already filter file names in the API.

\#1580

Adding Import/Export functionality

\#1660

send version info to stdout

\#1566

The pins API is unified and exposes only addresses of pinned roots without exposinge the associated counter to the users.

\#2433

this PR simplifies the loadsave abstraction to remove state like encryption and mode. those are now injected through a pipeline factory. It also handles more explicitly the cases where a readonly loadsaver is needed, so that readonly operations on manifest can happen. this is now possible by requesting explicitly a read-only loadsaver.

also adds missing unit test coverage and documentation.

\#1611

As per #1084 we need to return http response code 201 every time we create a new resource.

\#2545

The io/ioutil package has been deprecated in Go 1.16 (See https://golang.org/doc/go1.16#ioutil). This PR replaces the existing io/ioutil functions with their new definitions in io and os packages.

\#1490

This PR is putting back bmt under swarm. This package implements a crucial component of swarm which in turn is only relevant for swarm, hence its natural place is under bee.

The PR does not only migrate the existing repo in https://github.com/ethersphere/bmt to pkg/bmt, but also

simplify and clean up code
remove unused/useless function
reorganise tree and hasher fields
get rid of locking completely (was originally, but got corrupted on the way).. also tested with -race
get rid of section on hasher, use and reuse buffer on tree

\#2481

SWA-01-010 WP2: Unbounded recursion in file joiner (Medium)

While reviewing the file joiner implementation, it was found that the code allows for an
unbounded recursive call.

\#1600

Adding counter for error codes that are returned eg:

```
bee_api_reponse_error_codes{code="404",method="GET"} 3
```

\#1616

Rename address to reference in pinning related code

\#1962

api shutdown timeout is decreased, since shutting down with running websockets is not critical, while slowing down the rest of the process is

\#1406

pipeline: add special case for empty file

\#1659

cmd: explicitly discard returned values from winlogs